fix batch_completion
